move cli args and parsing to separate util to remove dependencies on entrypoint
